split BoardStateMachine into several input handler files
