+++ src/BoardStateMachine/BoardStateMachineUtils/index.ts
@@ -1,0 +1,9 @@
+import connectItem from './connectItem';
+import createItem from './createItem';
+import disconnectItem from './disconnectItem';
+import updateLineConnections from './updateLineConnections';
+
+// BoardStateMachine util functions are not pure functions like the ones in the src/utils folder
+// these import the store to check state and dispatch actions
+
+export { connectItem, createItem, disconnectItem, updateLineConnections };
