+++ src/BoardStateMachine/userInputHandlers/index.ts
@@ -1,0 +1,6 @@
+import handleMouseDown from './handleMouseDown';
+import handleMouseMove from './handleMouseMove';
+import handleMouseUp from './handleMouseUp';
+import handleMouseWheel from './handleMouseWheel';
+
+export { handleMouseDown, handleMouseMove, handleMouseUp, handleMouseWheel };
